requirements.txt
django==3.2.12
djongo==1.3.6
pymongo==3.12.3

# Native GUI Wrapper
pywebview>=4.0.0
